Changed URL for GAE SDK download
Changed the URL and linktext for the GAE SDK download link. I followed the instructions and downloaded and installed the SDK from the original link, but that installs only the SDK and not the Launcher (GUI). Trying to follow the instructions with that version or flavour of the SDK is impossible.